requirements.txt
diffusers==0.20.0
transformers==4.46.2
accelerate==0.26.0
huggingface-hub==0.25.0
einops==0.3.0
pytorch-lightning==1.4.2
torchmetrics==0.6.0
kornia==0.6
scipy
openai
opencv-python
safetensors
pyyaml
pydantic
dadaptation
lion-pytorch
prodigyopt
omegaconf
invisible-watermark==0.2.0
